In phone model change to Big Integer field

+++ src/dashboard/models.py
@@ -6,7 +6,7 @@ class UserProfile(models.Model):
     user_name = models.OneToOneField(User, on_delete=models.CASCADE)
     city = models.CharField(max_length=60, null=True)
     post_code = models.CharField(max_length=60, null=True)
-    phone_number = models.IntegerField(null=True, validators=[MaxValueValidator(99999999999999),MinValueValidator(100000000)])
+    phone_number = models.BigIntegerField(null=True, validators=[MaxValueValidator(99999999999999),MinValueValidator(100000000)])
     is_aider = models.BooleanField(blank=False, null=False, default=False)
 
     def __str__(self):
